GCS_MAVLink: the routing should only forward a message once per interface
On the previous code, when multiple routes shared the same interface a msg could be forwarded multiple times though the same port.

This fixes it by looping by all interfaces, and then looping thought the routes to check if any maches that interface. The alternative would be to loop thought routs and keep a vector of which interfaces the msg was forwarded to.